Refactor inventory application
