+++ start/src/main/java/io/openliberty/guides/inventory/InventoryResource.java
@@ -40,12 +40,16 @@ public class InventoryResource {
   @Path("{hostname}")
   @Produces(MediaType.APPLICATION_JSON)
   public Response getPropertiesForHost(@PathParam("hostname") String hostname) {
+    // Get properties
     Properties props = manager.get(hostname, 9080);
     if (props == null) {
       return Response.status(Response.Status.NOT_FOUND)
                      .entity("ERROR: Unknown hostname or the resource may not be running on the host machine")
                      .build();
     }
+
+    // Add properties to inventory
+    manager.add(hostname, props);
     return Response.ok(props).build();
   }
